Working config.json (and manual price updates?)
The configuration settings are properly being read/saved and are
implemented into the functionality of the updates and the alerts. The
enable/disable needs to be revised, but it works very well otherwise.
Also, manual price updates (on prices that aren’t just on the interest
list) magically started working… not sure exactly what I did but
they’re functional now.