'Show' functionality
Added a few keywords to the ‘show’ command and added functionality for
showing whether updates/alerts are enabled/disabled. Also changed some
of the grammar mistakes that were present along with some of the
verbiage of the bot. Everything is now working properly! (The only
thing that might be “if-y” now is the “show me the price of COIN”
command)